refactor: moved site_data_dir and get_asset_path to dirs.rs
